파일 분할 및 최종 수정
